Feature: Added ability to delete a To Do item
Loop through every To-Do item and show a delete button
Pressing the delete button sends a request that includes which to-do item to delete
The controller takes the user input, and notifies the models to delete the To-Do object by ID
On successful deletion by the models, the controller notifies the view to refresh the page and redirect to homepage, showing a fresh fetch of all To-Do items.

Create function in HTML view was also updated to include the completed checkbox and the delete button.